breaking: remove SecurityService interface
