--- src/api/KeycloakSecurityServiceOptions.ts
@@ -1,9 +1,0 @@
-export interface Logger {
-  info(...args: any[]): void
-  error(...args: any[]): void
-}
-
-export interface KeycloakSecurityServiceOptions {
-  log: Logger,
-  keycloak: any
-}
