+++ src/api/KeycloakSubscriptionHandlerOptions.ts
@@ -1,0 +1,3 @@
+export interface KeycloakSubscriptionHandlerOptions {
+  keycloak: any
+}
