+++ src/api/index.ts
@@ -1,7 +1,3 @@
-export * from './SecurityService'
-export * from './DefaultSecurityService'
 export * from './AuthContextProvider'
-export * from './SecurityService'
-export * from './ApplyAuthMiddlewareOptions'
-export * from './KeycloakSecurityServiceOptions'
+export * from './KeycloakSubscriptionHandlerOptions'
 export * from './SchemaDirectives'
