+++ src/directives/schemaDirectiveVisitors.ts
@@ -1,8 +1,26 @@
 import { defaultFieldResolver, GraphQLSchema } from 'graphql'
 import { SchemaDirectiveVisitor } from 'graphql-tools'
 import Joi from 'joi'
-import { directiveResolvers } from './directiveResolvers'
+import { auth, hasRole } from './directiveResolvers'
 import { VisitableSchemaType } from 'graphql-tools/dist/schemaVisitor'
+
+export class AuthDirective extends SchemaDirectiveVisitor {
+
+  constructor (config: {
+      name: string
+      visitedType: VisitableSchemaType
+      schema: GraphQLSchema
+      context: { [key: string]: any }
+    }) {
+    // see https://github.com/apollographql/graphql-tools/issues/837
+    super(config as any)
+  }
+
+  public visitFieldDefinition (field: any) {
+    const { resolve = defaultFieldResolver } = field
+    field.resolve = auth(resolve)
+  }
+}
 
 export class HasRoleDirective extends SchemaDirectiveVisitor {
 
@@ -26,7 +44,7 @@ export class HasRoleDirective extends SchemaDirectiveVisitor {
 
     const { roles } = value
 
-    field.resolve = directiveResolvers.hasRole(roles)(resolve)
+    field.resolve = hasRole(roles)(resolve)
   }
 
   public validateArgs () {
