+++ src/index.ts
@@ -1,2 +1,4 @@
-export * from './KeycloakSecurityService'
+export * from './KeycloakSubscriptionHandler'
+export * from './AuthContextProvider'
+export * from './directives'
 export * from './api'
